requirements.txt
django
djangorestframework
pillow
psycopg2
pygments
django-filter
django-guardian
markdown
django-cors-headers